% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{clusters_km}
\alias{clusters_km}
\title{clustering results of the k-mean algorithm}
\usage{
clusters_km(x, k = 2)
}
\arguments{
\item{x}{matrix of data  (dim 1: samples (must be equal to dim 1 of X), dim 2: attributes (must be equal to dim 2 of X))}

\item{k}{number of clusters ( this version considers 2 clusters )}
}
\value{
sum of square (SS) of clustring
}
\description{
clusters data into two clusters. This functionis uses the \code{kmeans} function to cluster the data and exports the clustering results as well as the sum of square (SS) of clustering using the Euclidian distance.
}
\examples{
{
X=rbind(matrix(rnorm(1000*2 ,4,.1),1000,2),matrix(rnorm(1000*2, 3, 0.2),1000,2))
M <- X[sample(nrow(X), 2),]
clusters_km(X,2)
}
}
