% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{clusters_MajKm}
\alias{clusters_MajKm}
\title{clustering results of the majorized k-mean algorithm}
\arguments{
\item{x}{matrix of data  (dim 1: samples (must be equal to dim 1 of X), dim 2: attributes (must be equal to dim 2 of X))}

\item{k}{number of clusters ( this version considers 2 clusters )}

\item{La}{the tunnung parameter}
}
\value{
sum of square (SS) of clustring and the 'delta' (difference of two successive majorization function).
}
\description{
clusters data into two clusters with a majorization k-means This functionis use a hybrid of the k-means and the majorizaion-minimazation method to cluster the data and exports the clustering results as well as the sum of square (SS) of clustering
}
\examples{
{
X=rbind(matrix(rnorm(1000*2 ,4,.1),1000,2),matrix(rnorm(1000*2, 3, 0.2),1000,2))
M <- X[sample(nrow(X), 2),]
clusters_MajKm(X,2, 0.5)
}
}
