% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{Euclid}
\alias{Euclid}
\title{Euclidian distance}
\usage{
Euclid(x, mu)
}
\arguments{
\item{x}{matrix of data  (dim 1: samples (must be equal to dim 1 of X), dim 2: attributes (must be equal to dim 2 of X))}

\item{mu}{initial seleted centroids (randomly or another method).}
}
\value{
Euclidian distance between two points.
}
\description{
Calculates the Euclidian distance between points. This function can use in \code{kmeans} function to do the clustering procedure using the Euclidian distance.
}
\examples{
{
X=rbind(matrix(rnorm(1000*2 ,4,.1),1000,2),matrix(rnorm(1000*2, 3, 0.2),1000,2))
M <- X[sample(nrow(X), 2),]
Euclid(X,M)
}
}
