% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{kmeans}
\alias{kmeans}
\title{k-means function}
\arguments{
\item{x}{matrix of data  (dim 1: samples (must be equal to dim 1 of X), dim 2: attributes (must be equal to dim 2 of X))}

\item{centers}{initial seleted centroids (randomly or another method)}

\item{distFun}{function (in this package the distance is Euclidian)}

\item{nItter}{Number of itteration function}
}
\value{
clustered results based on k-means methods.
}
\description{
k-means algorithm in clustering. This function export the clustered results based on one replication of the k-means method
}
\examples{
{
X=rbind(matrix(rnorm(1000*2 ,4,.1),1000,2),matrix(rnorm(1000*2, 3, 0.2),1000,2))
M <- X[sample(nrow(X), 2),]
kmeans(X,M, Euclid, 4)
}
}
